reworked utf8 library
